website/requirements.txt
Django==2.2
django-ckeditor==6.1.0
django-js-asset==1.2.2
djangorestframework==3.13.1
Pillow==8.3.0
pytz==2021.1
sqlparse==0.4.1
